Adding Support for systemd. Issue #16.
Install script tries to detect if systemd is present, and ask for
confirmation. This way we can support 14.04 for some more time.

For systemd I opted to use a wrapper script put in /opt/alfresco, this
will make it much easier for users to find the JAVA_OPTS settings.
This script can be called with alfresco-service.sh start/stop but will
actually use systemd;     sudo systemctl start|stop alfresco.service
This way it is kept in control of the systemd deamon.